Update to match 68k.
lap:make-unconditional-branch returns an instruction sequence rather
than a single instruction because on machines with branch delay slots
and no nullification, a NOP must appear as well.

Comment out uses of expt, since compiler declarations don't work with
compile-by-procedures?

Add a test on endianness to expand-fields so that the same macros can
be used for all RISC machines.