Added LOAD-LATEST and FASLOAD-LATEST to load the most recently modified version among the LOAD/DEFAULT-TYPES.
Added a variable LOAD/DEFAULT-FIND-PATHNAME-WITH-TYPE which may be set
to a procedure which tells load how to decide which file type to use.
The procedure accepts a pathname and a list of file type strings
(usually this will be LOAD/DEFAULT-TYPES) and should return the
complete pathname for load to use.