When the continuation parser makes a transition from interpreted to compiled frames or vice-versa, the history subproblem must be stepped. The reasons for this are that:
1. When interpreted code is called from compiled code, a new history
subproblem is pushed.  This subproblem must be discarded when making
the transition back to the compiled code stack frame.

2. When compiled code is called from interpreted code, the existing
history subproblem is preserved for use if the compiled code
tail-recurses back into interpreted code.  This subproblem must be
discarded when making the transition back to the interpreted code
stack frame.