Eliminate use of INTERRUPTABLE_EXTENT on read and write system calls. This doesn't work as the system call might finish and then check for signals just before returning with a valid answer (and thereby losing data).
This code should work OK for 4.3BSD, POSIX, and SYSV systems.
Unfortunately there's no good way to make it work for 4.2BSD -- which
probably means some hairy kludge will be needed to get ^G interrupts
to work on that system.  This will require identifying when we go into
an input wait on the console.