Add support for special keys, including function keys and arrow keys, with bucky bits.
A key is either a CHAR or a SPECIAL-KEY.  SPECIAL-KEYs can be compared
using EQ?.

Support all the keys named in /usr/include/X11/keysym.h, using the
names X gives them (with few exceptions).

Rename most procedures for dealing with keys from "-char" to "-key" to
prevent confusion.

Requires microcode 11.91 or later because of changes to the X keyboard
event structure.