Fix various bugs with evaluation commands:
Grab the transcript buffer (and whether it's enabled) BEFORE
beginning an evaluation, because, for example, the output from the
debugger Z command would go into a completely random place (window
and buffer).

Make sure that buffers for which the transcript buffer is not enabled
do not output values to the transcript buffer.