Make RETURN-TO and RETURN-FROM evaluate the previous expression instead of prompting for an expression to evaluate.
Provide evaluation commands that do and don't evaluate in the dynamic
state of the continuation.  Commands that evaluate in the dynamic
state of the continuation will have their output appear wherever the
program would have sent its output had no error occured, and fluid
variables will be bound as they would appear in the program.
Commands that do not evaluate in the dynamic state of the
continuation will have their output appear in the debugger buffer,
and fluid variables will be bound as they would upon leaving the
continuation.

Rearrange some key bindings.

Improve handling of restarts.