This microcode cannot be used with Edwin versions prior to 3.65.
The implementation of UX_select has been changed to return to the
caller when a Scheme interrupt needs servicing, rather than just
servicing the interrupt and going back into the select.  This gives
Scheme code a chance to run and do something useful.

In addition, the primitives CHANNEL-SELECT-THEN-READ and
X-DISPLAY-PROCESS-EVENTS have been changed to return different event
codes for the "other output" and "process status change" events.  This
allows the event-handling code to go directly to the handler, rather
than trying to figure out which one occurred.