COMINT-LINE-START was matching COMINT-PROMPT-REGEXP against the remainder of the buffer when it is only necessary to match against the remainder of the line.  This occasionally caused the primitive re-match-buffer to signal an anonymous error.  This patch will make the anonymous error extremely unlikely.
The anonymous error indicates a stack-overflow condition that is
difficult to fix properly.  However, the primitive can be modified to
use heuristics that will eliminate the error; the cost is that cases
that would have signalled the error will instead return a non-maximal
match or a non-match in cases where it would have returned a maximal
match had the stack been deep enough.