A few little changes, plus,
Change conditionalization of the open-coding of floating-point
primitives.  It was previously done statically when the compiler was
built.  It is now done at the point of the call, so the switch can be
meaningfully fluid-let around a compilation.

Add the unsafe open-coding of integer->char.
Add a couple of optimizations to plus-fixnum and minus-fixnum.