Change DOS timer interrupt behavior:
- Low level timer interrupt sets the "Global GC" interrupt but.

- The Global GC interrupt handler (a primitive) tracks user timer
requests and sets the timer interrupt bit when appropriate.  It also
invokes consume_typeahead to look for keyboard interrupt characters.

Keyboard interrupt behavior changes:

- The interactive handler is invoked from the consume_typeahead
routine, not when Scheme is trying to map the interrupt character.
In this way, its range of actions is greater.

- Add reset option and hard attention keys.

Keyboard interrupt handler:

- Implement X32 version (does not work under DPMI).

- Rewrite DPMI handler to use masm-assembled code instead of
hand-assembled code.  This makes it considerably easier to modify.