Change DYNAMIC-WIND (and FLUID-LET) so that they record the state transition *after* the "before" thunk is executed.  This guarantees that both the "before" and "after" thunks execute in the dynamic state outside of the DYNAMIC-WIND call.
**** NB: ****
The particular change I made is buggy: I only changed how the state
transition is performed in the normal case.  If you throw out from the
"during" thunk and then throw back in, the "before" thunk will be
executed after its state transition is recorded.  In order to fix this
properly, the state space structure must be extended to include
information indicating whether the transition should be recorded
before or after the transition thunk.