These changes require microcode version 11.135 or later.
Change the startup sequence for X windows.  It turns out that there is
a poorly documented constraint which says that the client may not draw
on a newly-created window until the first Expose event for the window
is received.  These changes implement this constraint (see the source
code comments for details).  Additionally, the startup sequence now
guarantees that the event mask and Edwin's data structures are
properly initialized before the window is mapped, guaranteeing that by
the time the first event arrives, everything is ready for it.