These changes require microcode version 11.145 or later.
Reimplement handling of temporary files to update list in fixed
objects vector; this list tells the microcode that the files should be
deleted when Scheme is killed.

Additionally, change the handling of ".bci" file expansion to use the
new temporary file mechanism.  The expander now accepts a third value
for *SAVE-UNCOMPRESSED-FILES?* which says to expand the file into a
temporary file that is deleted (by the GC) after it has not been used
in a while, or when Scheme is killed.  This new option is now the
default, and the timeout for these temporary files defaults to five
minutes.