Changed C_call_scheme to preserve and restore `last_return_code'. This fixes a bug which manifests as an access violation during GC.
The GC bug happened because the incorrect value of `last_return_code'
left after the recursive interpretation caused a negative stack
segment size to be computed.  When this was boxed as an unsigned value
it became an illegal pointer object.  Note that this manifestation
would not have happened with the new tagging scheme.