Changing WITHOUT-INTERRUPTS to WITH-EDITOR-INTERRUPTS-DISABLED was not the right thing.
WITH-EDITOR-INTERRUPTS-DISABLED is required in the running editor but
the first buffer is created before the editor structure exists, and
variables may be set as early as initial file load time (prior to
dumping the band) -- e.g vc.scm.

WITHOUT-EDITOR-INTERRUPTS choses to do either a WITHOUT-INTERRUPTS or
a WITH-EDITOR-INTERRUPTS-DISABLED (confused?).  Perhaps
WITH-EDITOR-INTERRUPTS-DISABLED ought to have this behaviour.