Fixed edwin.ldr.  Some files were loaded twice.
Changed the selection of a default display type.

If a display type is requested but is not available then an error is
generated (the code used not to check).

There is no longer a preference for the display type 'CONSOLE, so the
list of defined types is always searched.  The initialization sequence
in edwin.ldr is tterm before (xterm or win32), which puts the
graphical display types ahead of the console in the display types
list, and this places them at a higher desirability.

Doing this revealed that the win32 availablility predicate was
implemented incorrectly.