Added the profile information abstraction.
The base abstraction of the profile info keys and tags should probably
go in the runtime.

A reporting package should be created as a load option.

The important parts to the compiler should stay here.  These parts
track the profile-data declarationd and compute the information that
is inserted in the compiled code block.