Added debugging environment info.
Because cleanup and lamlift do not yet understand the debugging info,
this does not help much.  Lamlift might not be the problem since
breaks the program into lots of little pieces which end up by being
put back together by simplify & cleanup, so these could lose the
information.

I tested it on a small example with

  (set! *phases-to-omit* '(lamlift/top-level/1 lamlift/top-level/2))

and it seems to produce the right expressions in the debugging environment.

NOTE: bigger examples do not work because closconv expects bindings to
be either operators or operands but not both (i.e. it assumes that
lamlift has `split' the uses).

NOTE: the processing of the environment is wrong at LAMBDAs: currently
the references are processed in the environment OUTSIDE the LAMBDA,
but the debugging information is supposed to be for INSIDE the lambda.

This is going to be a general problem with LAMBDAs.