Added a patch to ensure that out-of-line hooks are not called with a constructed continuation.
	  (call 'hook (call %make-stack-closure ...) x y)
	=>
	  (call %invoke-continuation
	        (call %make-stack-closure ...)
		(call 'hook '#F x y))

This is required because out-of-line hooks have to generate their own
continuation to do the restoring, and the code merges with a control
flow path from somewhere prior to the call so that this continuation
is part of the same rgraph, i.e. the RTL graph sblock for an
INVOKE-SPECIAL-PRIMITIVE is connected to the bblock for the
continuation.  This means that RTL pseudo register value classes must
agree across this boundary, which is not so if we generate one from
%make-stack-closure.

Also tidies a few things.