Removed %vector-index expressions and the indexify phase.  The index computation is now implicit and handled by rtlgen and dbgred.
Looking at the structure of some large programs (e.g. symb in
nscmutils) revealed that up to 32% of the text of a program was
%vector-index expressions.  These get copied in every pass after
closconv/2.  The `elegance' is not worth the performance hit.  About
30% of the remaining text is %stack-closure-ref expressions.  The
overhead could be reduced by introducing another KMP special form
(which would also mean that the DBG info would not need to compress
the expressions, saving time).