Added EQUAL? and CEILING, FLOOR, ROUND, TRUNCATE to known global operators.
Changed expansions for FIX:=, FIX:ZERO? to NOT use EQ? as this hides
type info (i.e. that argument is a fixnum) from the compiler.

Moved CELL?, FLO:FLONUM? to gconst.scm to use `native' expansion.
There is no speed benefit to interpreted code and no benefit to
compiler, so it is clearer to leave it in.  It is a pity that there is
no primitive VECTOR?