Added code to deal with primitives called with the wrong number of arguments.  A warning is issued.  The primitive is called with %internal-apply to ensure correct behaviour at run time.
Note: Perhaps this processing should go later, in case something
clever finds more primitives to substitute.  For non open-coded
primitives, this can be done as late a compat.  The problem with this
is that after CPS conversion, it is impossible to convert an open
coded primitive into a subproblem call.  One way to solve this is not
open code NO primitives, only cookie calls.