The following change in how WIN32 graphics devices are flushed gives non-flickering update almost like double-buffering for programs which have to draw multiple frames from scratch.
If graphics buffering is enabled, then a GRAPHICS-FLUSH operation
comes from the user's program.  The change is to synchronously redraw
the screen in this (i.e. buffered+flush) case.  Subsequent operations
can clear and draw without an asynchronous redraw operation copying
the subsequent partial image to the screen.

This is not quite double-buffering since forced redisplay during the
buffered drawing (for example, uncovering part of the window) will
cause the partly draw backing bitmap to be copied to the screen.

User programs with buffering and frequent explicit flushes are likely
to be slower.