This version of Edwin requires microcode 11.162 or later if used under Win32.
Extensive changes to use new single input queue, and to implement
event-stream previewer.  (The single input queue was required for
proper implementation of the previewer.)  With this change, the Win32
platform has the ability to abort out of any Edwin command, as has
been true of all other platforms for years.

I also took the opportunity to considerably clean up the event code.