Change arguments to OTHER-SCREEN so that only first is required.
Eliminate OTHER-SCREEN?, replacing occurrences with OTHER-SCREEN.

Change meaning of SCREEN-VISIBLE? to mean "not unmapped"; this is
almost always what was meant.

Fix bug in SELECT-WINDOW: when selecting the selected window of
another frame, must select the frame.

Fix SELECT-BUFFER-OTHER-SCREEN to select the frame as well as the
buffer if reusing an existing frame.