New operation CONTAINER-URL-CONTENTS.
Reimplemented URL-IS-SELECTABLE? for file folders.  New implementation
probes the file to determine if it is a known type.

File-folder completion now only considers files of known type and
directories.  It also doesn't consider "." and ".." directories.

Code that mapped IMAP heirarchy delimiters was broken, although this
caused no practical consequences.  It has been reimplemented to make
it both correct and simpler.

IMAP-folder completion used to probe subfolders of a folder to
determine if a folder should have a "/" at the end.  This was wasteful
of network bandwidth and had no practical consequences, so it has been
changed to not do this.